refactor(package): move authentication into data
