+++ app/src/main/java/com/example/ezvault/data/authentication/authentication/FirebaseAuthenticationStrategy.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.authentication.authentication;
+package com.example.ezvault.data.authentication.authentication;
 
 import com.example.ezvault.data.database.FirebaseBundle;
 import com.example.ezvault.model.User;
